Adding syslog support.
Add support for sending logs directly to syslog using
the python logging libraries "SysLogHander"

The change introduces a "logtype" feature whereby the user can
choose to log to a file, syslog or stderr

The existing functionality is retained, in as much as simply
specifying a logfile will cause it to act as if you also specified
the "file" logtype.  Also if "file" logtype is specified but no file is
specified it will default to logging to stderr.