Produce valid JSON for all input with the logstash log format.
The old method of constructing the log messages for the logstash
format produced invalid JSON for certain input. We avoid this by
instead using Python's standard json library for producing the JSON
string. As a bonus this allows us to drop several json.dumps() calls
in the log statements themselves. This addresses issue #248.