Fix #309
This corrects an issue where if a zero value were provided to `--newer-than` or `--older-than` it was caught by a boolean "is the variable present" test.

This also cleans up the get_target_month method with the addition of the month_bump method.

Unit and integration tests have been updated to catch these items.

fixes #309