Proposed fix for #348
Add a warning and a countdown (which allows this to still work in
a cron situation, without the need for crazy dialog, Y/N, etc.) in
the event you are filtering by `--timestamp` and not using the
`--older-than` or `--newer-than` flags.  This gives users 10 seconds
to "bow out" before it will act.

Update tests to test this functionality (very passively).