Convert settings to functions
All constants from `settings` have been replaced with a function.  This fixes the issue that required copy/deepcopy in the first place.  Values shouldn't be stomped on in between runs.  A test derived from a user case is employed to prove functionality.

Change defaults to dir structure for python 3.x.  This addresses Python 3.x issues with local module import

Also skip a test for 5.0.0-alpha4 + as it no longer allows the end user to force the creation_date setting.

Catch errors, failures, and skips. Allow for the known skipped test right now.

fixes #687