Add configuration validation, set defaults.
This also adds the voluptuous python module to the list of requirements.

Update defaults in repomgrcli.py to have hard-coded defaults (for now).

Massive refactor for config validation

These changes include all changes necessary to enable #674

This also includes changes to documentation discovered to be deficient when referencing changes.

fixes #674