Add alias filtertype
This feature allows the user to filter an IndexList based on whether the indices are associated with one or more aliases.

This commit also fixes the absence of configuration validation testing for whether filters were appropriate to the list type, i.e. a SnapshotList or IndexList.

Documentation added, tested, and the appropriate tests also added.

fixes #634