Preparation for 5.4 release
Replace the `field_stats` API call with a query, and now we can support Elasticsearch 6.0!

Fix and add missing tests

Somehow I missed indexsettings and shrink, and needed to restore them from TimeMachine.

This replaces them, and corrects the field_stats tests to work with the new query based version

Improve docs

Fix the docs a bit to show 6.0 compatibility.