Fix `ignore_empty_list` with Alias action
If `ignore_empty_list` is `true`, Curator should respect that and not raise an `ActionError`

Did some virtualenv testing and found that not specifying `six` in requirements.txt resulted in execution errors.

Restored the `before_install` block to `.travis.yml`.  Before this, it worked for _each_ run, _except_ for the elasticsearch 6.2.4 tests, which I cannot explain.  Best to keep it, since it works.

fixes #1209