Update get_client and cli.py client logic
This is an attempt to make more clear what's going on behind the
scenes in issues like #1434. This may or may not fix said issues,
but it should improve visibility enough to take additional steps,
if needed. Marking the issue as closed, as any further issues will
be different.

Fixes #1434