Add username and password to client configuration
Deprecate http_auth in favor of username and password.

Lots of code clean-up. `get_client` was broken out into separate functions to clean it up, as it was enormous.

Ignore module import monkey-patch for Python 2

Python2 is no longer going to be supported, either by Python or by Curator after 2019-12-31