[docs] clarification on prefix and suffix kinds (#1558)
Clarify that values of `prefix` and `suffix` filter kinds are not taken literally but rather a regexp is constructed from them.
This is important difference if index names are using dots (or other special characters).
For example, with  prefix of `value: project.prod-a.` the index `project.prod-admin-portal.2020.10.15` is matched. By reading the documentation this is unexpected as the impression is that prefix and suffix should be used literally.