Fix unescaped characater error in docstring
There is an unescaped character in the docstring of indexlist module which has
been deprecated since Python 3.6. Ref:
https://docs.python.org/3/whatsnew/3.6.html#deprecated-python-behavior

This causes an issue during pytest's test collection phase: "SyntaxError:
invalid escape sequence '\d'" when running tests on modules that import curator
as a library on later versions of Python.

In popular python linters e.g. ruff and flake8 this violates rule W605. Ref:
https://docs.astral.sh/ruff/rules/invalid-escape-sequence/

Upon linting the entire source code, this is the only file that has this rule
violation. This commit fixes this error.