Adding deepfreeze group and multiple actions
This update alters the click configuration so that deepfreeze is a group
and each action is a command. I've coded do_<action> methods in the
actions file, but these aren't functional yet. I need to get with Aaron
to see how the plumbing is done so I don't do something weird. There's
no command currently set up with sub-commands, so I don't have a pattern
to go from.