Improvements to curator/helpers/utils.py
There was a shortcoming in the snapshot restore action class. It was relying on the user to
provide the exact index names to restore. The restore API call to Elasticsearch allows for
multi-target syntax to select and de-select indices by comma-separated patterns. While not
expressly allowed, it was possible to use. The problem was that indices could not be properly
matched and verified if patterns rather than exact names were used. Three functions were added
to ``helpers.utils`` to make this work properly: ``multitarget_match``, ``multitarget_fix``,
and ``regex_loop``.

The ``Restore`` class now calls ``multitarget_match`` in the ``_get_expected_output`` method
to get the names from the snapshot list object. This now works with patterns, and the tests
have been updated to ensure this.