Fix singleton show_indices
Addresses #1775

Fixed some Pylance errors with re-doing how CLIAction.get_list_object() works (return the value now instead of setting the attribute), and instead of setting self.list_object = None in the early lines of `__init__`, move it to the end so all of the dependencies needed by get_list_object() are addressed.

Added a bit of typing to object_class.py as well.

In show.py, the missing step was

```python
action.list_object.get_index_stats()
```

This was missed when the IndexList class was updated to change how and when stats were collected.