Add fix for metadata repair
Detect and fix situation where a thaw request is submitted, acted upon
by AWS, but ignored by the requestor. If check-status is run after the
data is refrozen by AWS, this detects that and fixes the metadata to
show the request as being refrozen so it doesn't languish as a pending
request.